move code from lib/addons to manager-api, reference manager-api or preview-api instead of lib/addons
