+++ code/addons/a11y/src/manager.test.tsx
@@ -1,4 +1,3 @@
-import { addons } from '@storybook/addons';
 import * as api from '@storybook/manager-api';
 import { PANEL_ID } from './constants';
 import './manager';
@@ -7,7 +6,7 @@ jest.mock('@storybook/manager-api');
 jest.mock('@storybook/addons');
 const mockedApi = api as unknown as jest.Mocked<api.API>;
 mockedApi.getAddonState = jest.fn();
-const mockedAddons = addons as jest.Mocked<typeof addons>;
+const mockedAddons = api.addons as jest.Mocked<typeof api.addons>;
 const registrationImpl = mockedAddons.register.mock.calls[0][1];
 
 describe('A11yManager', () => {
