+++ code/addons/controls/src/manager.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { addons, types } from '@storybook/addons';
+import { addons, types, type API, useArgTypes } from '@storybook/manager-api';
 import { AddonPanel } from '@storybook/components';
-import { type API, useArgTypes } from '@storybook/manager-api';
 import { ControlsPanel } from './ControlsPanel';
 import { ADDON_ID, PARAM_KEY } from './constants';
