+++ code/addons/docs/docs/multiframework.md
@@ -112,6 +112,8 @@ This dynamic rendering is framework-specific, meaning it needs a custom implemen
 
 Let's take a look at the React framework implementation of `dynamic` snippets as a reference for implementing this feature in other frameworks:
 
+TODO: understanding this better, should probably import from `preview-api` instead of `addons`?
+
 ```tsx
 import { addons, StoryContext } from '@storybook/addons';
 import { SNIPPET_RENDERED } from '../../shared';
