+++ code/addons/interactions/src/manager.tsx
@@ -1,4 +1,4 @@
-import { addons, types } from '@storybook/addons';
+import { addons, types } from '@storybook/manager-api';
 
 import { ADDON_ID, PANEL_ID } from './constants';
 import { Panel } from './Panel';
