+++ code/addons/links/package.json
@@ -80,6 +80,7 @@
     "@storybook/client-logger": "7.0.0-alpha.53",
     "@storybook/core-events": "7.0.0-alpha.53",
     "@storybook/csf": "next",
+    "@storybook/manager-api": "7.0.0-alpha.53",
     "@storybook/preview-api": "7.0.0-alpha.53",
     "@storybook/router": "7.0.0-alpha.53",
     "@storybook/types": "7.0.0-alpha.53",
