+++ code/addons/measure/src/manager.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { addons, types } from '@storybook/addons';
+import { addons, types } from '@storybook/manager-api';
 
 import { ADDON_ID, TOOL_ID } from './constants';
 import { Tool } from './Tool';
