+++ code/addons/outline/src/manager.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { addons, types } from '@storybook/addons';
+import { addons, types } from '@storybook/manager-api';
 
 import { ADDON_ID } from './constants';
 import { OutlineSelector } from './OutlineSelector';
