+++ code/addons/storyshots/storyshots-core/src/api/index.ts
@@ -1,5 +1,5 @@
 import global from 'global';
-import { addons, mockChannel } from '@storybook/addons';
+import { addons, mockChannel } from '@storybook/preview-api';
 import ensureOptionsDefaults from './ensureOptionsDefaults';
 import snapshotsTests from './snapshotsTestsTemplate';
 import integrityTest from './integrityTestTemplate';
