+++ code/addons/storysource/src/manager.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { addons } from '@storybook/addons';
+import { addons } from '@storybook/manager-api';
 
 import { StoryPanel } from './StoryPanel';
 import { ADDON_ID, PANEL_ID } from './index';
