+++ code/lib/addons/src/storybook-channel-mock.ts
@@ -1,10 +1,1 @@
-import { Channel } from '@storybook/channels';
-
-export function mockChannel() {
-  const transport = {
-    setHandler: () => {},
-    send: () => {},
-  };
-
-  return new Channel({ transport });
-}
+export { mockChannel } from '@storybook/manager-api';
