+++ code/lib/manager-api/src/index.tsx
@@ -483,3 +483,8 @@ export function useArgTypes(): API_ArgTypes {
 }
 
 export { eventToShortcut, shortcutToHumanString, shortcutMatchesShortcut };
+
+export { addons } from './lib/addons';
+
+/* deprecated */
+export { mockChannel, types, type Addon, type AddonStore } from './lib/addons';
