+++ code/lib/manager-api/src/lib/storybook-channel-mock.ts
@@ -1,0 +1,10 @@
+import { Channel } from '@storybook/channels';
+
+export function mockChannel() {
+  const transport = {
+    setHandler: () => {},
+    send: () => {},
+  };
+
+  return new Channel({ transport });
+}
