+++ code/lib/preview-api/README.md
@@ -6,6 +6,7 @@ TODO write proper documentation of this package
 
 This package used to be multiple packages (they have been combined into this one):
 
+- `@storybook/addons` [read (old) docs](./README-addons.md)
 - `@storybook/client-api` [read (old) docs](./README-client-api.md)
 - `@storybook/core-client` [read (old) docs](./README-core-client.md)
 - `@storybook/preview-web` [read (old) docs](./README-preview-web.md)
