+++ code/ui/.storybook/manager.ts
@@ -1,4 +1,4 @@
-import { addons } from '@storybook/addons';
+import { addons } from '@storybook/manager-api';
 import startCase from 'lodash/startCase';
 
 addons.setConfig({
