+++ code/ui/.storybook/preview.tsx
@@ -10,10 +10,9 @@ import {
   styled,
   useTheme,
 } from '@storybook/theming';
-import { useArgs } from '@storybook/addons';
+import { useArgs, DocsContext } from '@storybook/preview-api';
 import { Symbols } from '@storybook/components';
 import type { PreviewWeb } from '@storybook/preview-api';
-import { DocsContext } from '@storybook/preview-api';
 import type { ReactRenderer } from '@storybook/react';
 import type { Channel } from '@storybook/channels';
