+++ code/ui/blocks/src/controls/Boolean.stories.tsx
@@ -1,7 +1,7 @@
 import { expect } from '@storybook/jest';
 import type { Meta, StoryObj } from '@storybook/react';
 import { within, fireEvent } from '@storybook/testing-library';
-import { addons } from '@storybook/addons';
+import { addons } from '@storybook/manager-api';
 import { RESET_STORY_ARGS, STORY_ARGS_UPDATED } from '@storybook/core-events';
 import { BooleanControl } from './Boolean';
