+++ code/ui/manager/src/components/preview/preview.tsx
@@ -2,9 +2,16 @@ import React, { Fragment, useMemo, useEffect, useRef, useState } from 'react';
 import { Helmet } from 'react-helmet-async';
 import global from 'global';
 
-import { type API, Consumer, type Combo, merge } from '@storybook/manager-api';
+import {
+  type API,
+  Consumer,
+  type Combo,
+  merge,
+  addons,
+  types,
+  type Addon,
+} from '@storybook/manager-api';
 import { PREVIEW_BUILDER_PROGRESS, SET_CURRENT_STORY } from '@storybook/core-events';
-import { addons, types, type Addon } from '@storybook/addons';
 
 import { Loader } from '@storybook/components';
 import { Location } from '@storybook/router';
