+++ code/ui/manager/src/components/preview/tools/copy.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import copy from 'copy-to-clipboard';
 import { getStoryHref, IconButton, Icons } from '@storybook/components';
 import { Consumer, type Combo } from '@storybook/manager-api';
-import type { Addon } from '@storybook/addons';
+import type { Addon } from '@storybook/manager-api';
 
 const { PREVIEW_URL, document } = global;
