+++ code/ui/manager/src/components/preview/tools/eject.tsx
@@ -2,7 +2,7 @@ import global from 'global';
 import React from 'react';
 import { getStoryHref, IconButton, Icons } from '@storybook/components';
 import { Consumer, type Combo } from '@storybook/manager-api';
-import type { Addon } from '@storybook/addons';
+import type { Addon } from '@storybook/manager-api';
 
 const { PREVIEW_URL } = global;
