+++ code/ui/manager/src/components/preview/tools/menu.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { IconButton, Icons, Separator } from '@storybook/components';
 import { Consumer, type Combo } from '@storybook/manager-api';
-import type { Addon } from '@storybook/addons';
+import type { Addon } from '@storybook/manager-api';
 
 const menuMapper = ({ api, state }: Combo) => ({
   isVisible: state.layout.showNav,
