+++ code/ui/manager/src/globals/exports.ts
@@ -200,11 +200,14 @@ export default {
     'Consumer',
     'ManagerContext',
     'Provider',
+    'addons',
     'combineParameters',
     'eventToShortcut',
     'merge',
+    'mockChannel',
     'shortcutMatchesShortcut',
     'shortcutToHumanString',
+    'types',
     'useAddonState',
     'useArgTypes',
     'useArgs',
@@ -218,11 +221,9 @@ export default {
     'useStorybookState',
   ],
   '@storybook/addons': [
-    'AddonStore',
     'HooksContext',
     'addons',
     'applyHooks',
-    'isSupportedType',
     'makeDecorator',
     'mockChannel',
     'types',
