+++ code/ui/manager/src/runtime.ts
@@ -1,8 +1,8 @@
 import global from 'global';
 
 import type { Channel } from '@storybook/channels';
-import type { AddonStore } from '@storybook/addons';
-import { addons } from '@storybook/addons';
+import type { AddonStore } from '@storybook/manager-api';
+import { addons } from '@storybook/manager-api';
 import type { Addon_Types, Addon_Config } from '@storybook/types';
 import * as postMessage from '@storybook/channel-postmessage';
 import * as webSocket from '@storybook/channel-websocket';
