+++ docs/snippets/common/storybook-manager-custom-theme.js.mdx
@@ -1,7 +1,7 @@
 ```js
 // .storybook/manager.js
 
-import { addons } from '@storybook/addons';
+import { addons } from '@storybook/manager-api';
 import yourTheme from './YourTheme';
 
 addons.setConfig({
