+++ docs/snippets/common/storybook-manager-dark-theme.js.mdx
@@ -1,7 +1,7 @@
 ```js
 // .storybook/manager.js
 
-import { addons } from '@storybook/addons';
+import { addons } from '@storybook/manager-api';
 import { themes } from '@storybook/theming';
 
 addons.setConfig({
