+++ docs/snippets/common/storybook-manager-render-label-stories.js.mdx
@@ -1,7 +1,7 @@
 ```js
 // .storybook/manager.js
 
-import { addons } from '@storybook/addons';
+import { addons } from '@storybook/manager-api';
 
 import startCase from 'lodash/startCase';
